refactor: move functions to parser utils
