+++ backend/editor/entries.py
@@ -5,7 +5,7 @@ import re
 import tempfile
 import urllib.request  # Sending requests
 
-from openfoodfacts_taxonomy_parser import normalizer  # Normalizing tags
+from openfoodfacts_taxonomy_parser import utils as parser_utils  # Normalizing tags
 from openfoodfacts_taxonomy_parser import parser  # Parser for taxonomies
 from openfoodfacts_taxonomy_parser import unparser  # Unparser for taxonomies
 
@@ -224,7 +224,7 @@ class TaxonomyGraph:
         """
         Helper function to check if a branch name is valid
         """
-        return normalizer.normalizing(self.branch_name, char="_") == self.branch_name
+        return parser_utils.normalizing(self.branch_name, char="_") == self.branch_name
 
     async def create_project(self, description):
         """
@@ -470,7 +470,9 @@ class TaxonomyGraph:
                     keys_language_code = keys.split("_", 1)[1]
                     normalised_value = []
                     for values in new_node_keys[keys]:
-                        normalised_value.append(normalizer.normalizing(values, keys_language_code))
+                        normalised_value.append(
+                            parser_utils.normalizing(values, keys_language_code)
+                        )
                     normalised_new_node_keys[keys] = normalised_value
                     normalised_new_node_keys["tags_ids_" + keys_language_code] = normalised_value
                 else:
@@ -556,7 +558,7 @@ class TaxonomyGraph:
         """
         # Escape special characters
         normalized_text = re.sub(r"[^A-Za-z0-9_]", r" ", text)
-        normalized_id_text = normalizer.normalizing(text)
+        normalized_id_text = parser_utils.normalizing(text)
 
         # If normalized text is empty, no searches are found
         if normalized_text.strip() == "":
